Work (incomplete) to store implicit taggings in the database.
Note: the 'untag' command has not yet been updated and so will leave implicit taggings in the database when a directory has an explicit tag removed.
Note: there is a bug in the 'merge' command where duplicate taggings can be inserted into the database in certain situations.

'delete', 'files', 'merge', 'stats', and 'tag' commands updated to handle implicit tags in the database.
Schema auto-generation updated with new columns.
Database upgrade script added to repository.
Updated logic for identifying files with tags to take advantage of the implicit taggings being in the database (should be a lot more performant).
Changes to storage layer and VFS to reflect implicit tagging changes.