Added support for different fingerprint algorithms.
The current algorithm, which is SHA256 except only sections of larger files are fingerprinted, is now called 'dynamic:SHA256' and remains the default.
Support has been added for 'SHA256' (irrespective of file size), 'MD5', 'dynamic:MD5' and 'symlinkTargetName'.
'SymlinkTargetName' has been added in the hope it will be sufficient to make use of TMSU work with Git Annex. Instead of calculating a fingerprint using a hash algorithm, instead it will look at the filename of a symbolic link target and use that. For regular files it will use the filename.