Fixed 'database locked' problem.
'database locked' was caused by querying outside a transaction before
starting a transaction. Apparently this locks the database in some kind
of implicit transaction. Who knew?

Also renamed the 'revision' column of 'version' table to 'patch' to use
the Semantic Versioning convention. If you happened to run TMSU in this
half hour window then you will need to drop the 'version' table so that
it can be recreated with the new name:

    sqlite3 /path/to/your/db
    drop table version