Bug fixes.
Removed the '-' argument to read subcommands from standard input as it
was a bit confusing, especially when subcommands like 'help' or 'vfs'
were used. Instead 'tag' (only) supports an argument of '-' to read the
file and list of tags to apply to it from standard input.

Subcommands now manage their own database transactions (for sanity).

Fixed bug with changes made via the virtual filesystem not being
committed.

Fixed up the Zsh completion.