Issue #27: VFS symbolic link sometimes wrong.
Code was parsing the wrong component of the filename as the file
identifier. Mostly VFS filenames are of the format 'name.3.ext', where
the file has an extension or 'name.3' where it does not (e.g. most
directories). The old code would first attempt to parse the ID from the
penultimate element before trying the last element. In the case of
numeric directories this meant the directory name was assumed to be the
ID.