Issue #39: Tag values are difficult to work with
Added support for tag-values in implications. This lets you put a
tag=value pair on either side of a tag implication.

This functionality very nearly completely works...except for the case
where you perform a greater-/less-than comparison, e.g.

    $ tmsu files "year > 2000"

This will not identify any files that are implicitly tagged with a year
bigger than 2000: I still need to implement this as it's quite fiddly.
However a simple equality works just fine:

    $ tmsu files "year == 2015"