Issue #54: Relax tag/value name restrictions.
Equals character can now be escaped in commands expecting a tag=value
pair.
Equals now coloured or escaped to differentiate it from literal equals
within a tag/value name. When colour is off it is shown escaped with a
backslash.
'imply' listing now shares colouring logic so this listing is now
coloured.