Makefile: dependencies and PHONY targets
The unit testing and integration testing target now depend on the
compilation target, which fixes a possible attempt to run tests before
the compilation finishes. This issue did not show up when `make compile`
(or `make dist`) was invoked before `make test`, but using the default
target with multiple jobs (e.g. `make -j 4`) could kick off a wrong
order of execution. This is now fixed.

Furthermore, all makefile targets that aren't files (in this case, all
targets) are made PHONY targets, see
[here](https://stackoverflow.com/questions/2145590/what-is-the-purpose-of-phony-in-a-makefile)
or
[here](https://www.gnu.org/software/make/manual/html_node/Phony-Targets.html)
for a reasoning.