Implement BASH completion
Known limitations:
1. Does not work correctly if there is 2 (or more) tags with spaces
in them and the parts before the spaces match (e.g. "space tag1"
and "space tag2");
2. Subcommand 'tag' is not completed correctly if options
'-f/--from/-t/--tags/-w/--where' are not last;
3. Subcommand 'untag' is not completed correctly if option
'-t/--tags' is not last;
4. Completion of subcommands 'files' and 'tag' (which can use
queries) does not show operators in queries because some of them
are not completed correctly (or at all) so they should be
entered by hand.